% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellDART_R.R
\name{pred_cellf_celldart}
\alias{pred_cellf_celldart}
\title{R wrap function to implement CellDART}
\usage{
pred_cellf_celldart(
  sp_data,
  sc_data,
  outdir = ".",
  sp_subset = FALSE,
  spot.cluster.name = "seurat_clusters",
  spot.cluster.of.interest = NULL,
  env.select = "conda",
  python.install = F,
  python_path = NULL,
  env.name = "CellDART",
  gpu = TRUE,
  metadata_celltype = "celltype",
  num_markers = 20,
  seed_num = 0,
  nmix = 8,
  npseudo = 20000,
  alpha = 0.6,
  alpha_lr = 5,
  emb_dim = 64,
  batch_size = 512,
  n_iterations = 3000,
  init_train_epoch = 10
)
}
\arguments{
\item{sp_data}{spatial data (Seurat object) to be used in predicting cell fraction: non-normalized raw data should be in 'counts' slot}

\item{sc_data}{single-cell data (Seurat object) to be used in making pseudospots: non-normalized raw data should be in 'counts' slot}

\item{outdir}{the directory to save output files (models and results) (default = '.')}

\item{sp_subset}{whether to subset spatial data and calculate for specific spot cluster (default = FALSE)}

\item{spot.cluster.name}{group name of the cluster used for subsetting spatial data (default: 'seurat_clusters')}

\item{spot.cluster.of.interest}{name of each spot clusters to be used (default: NULL)}

\item{env.select}{select between using reticulate virtual environment or conda environment (default: "conda")}

\item{python.install}{whether to automatically install python version 3.7.12}

\item{python_path}{path for the python 3.7. (default: NULL)
\itemize{
\item If NULL, python version 3.7.12 will be installed (valid for Linux)
\item If "current", python interpreter associated with current virtual env (ex: r-reticulate) will be used. (version should be 3.7)
}}

\item{env.name}{name of the virtual or conda environment to use for CellDART analysis (default: 'CellDART')}

\item{gpu}{check whether to use gpu (True) or not (False) (default = True)}

\item{metadata_celltype}{column name for single-cell annotation data in metadata (default: 'celltype')}

\item{num_markers}{number of selected marker genes in each cell-type (default = 20)}

\item{seed_num}{seed to be used in random sampling (default = 0)}

\item{nmix}{the number of cells sampled from single-cell data when making a pseudospot (default = 10)}

\item{npseudo}{a total number of pseudospots (default = 20000)}

\item{alpha}{loss weights of domain classifier to the source classifier (default = 0.6)}

\item{alpha_lr}{learning rate for the domain classifier (alpha_lr*0.001, default = 5)}

\item{emb_dim}{output size of dimensions for feature extractor (default = 64)}

\item{batch_size}{minibatch size for pseudospots and spatial data during the training (default = 512)}

\item{n_iterations}{iteration number for the adversarial learning (default = 3000)}

\item{init_train_epoch}{iteration number for the pre-training process (default = 10)}
}
\value{
spatial data (Seurat object) with predicted cell fraction in metadata (meta.data)
}
\description{
Cell type inference by domain adaptation of single-cell and spatial transcriptomic data
}
\examples{
Using conda environment (environment will be automatically installed in Linux distributions)
If using Windows, then install conda environment first and then run the function below with python.install = F
sp_data_cellf <- pred_cellf_celldart(sp_data, sc_data, outdir = '.',
                                     sp_subset=F, spot.cluster.name='seurat_clusters',
                                     spot.cluster.of.interest=NULL,
                                     env.select='conda',python.install=T,
                                     python_path=NULL, env.name='CellDART',
                                     gpu=TRUE, metadata_celltype='celltype',
                                     num_markers=20, seed_num=0,
                                     nmix=8, npseudo=20000, alpha=0.6,alpha_lr=5,
                                     emb_dim=64,batch_size=512,n_iterations=3000, init_train_epoch=10)

Using virtual environment (environment will be automatically installed in Linux distributions)
Not recommended for Windows
sp_data_cellf <- pred_cellf_celldart(sp_data, sc_data, outdir = '.',
                                     sp_subset=F, spot.cluster.name='seurat_clusters',
                                     spot.cluster.of.interest=NULL,
                                     env.select='virtual',python.install=T,
                                     python_path=NULL, env.name='CellDART',
                                     gpu=TRUE, metadata_celltype='celltype',
                                     num_markers=20, seed_num=0,
                                     nmix=8, npseudo=20000, alpha=0.6,alpha_lr=5,
                                     emb_dim=64,batch_size=512,n_iterations=3000, init_train_epoch=10)
}
